+++ packages/react-server/src/features/request-context/server.ts
@@ -1,20 +1,20 @@
-import { RequestCookies, ResponseCookies } from "@edge-runtime/cookies";
+import { createNextCookies } from "../next/cookie";
 import type { RevalidationType } from "../server-component/utils";
 import { createContextStorage } from "./utils";
 
 const requestContextStorage = createContextStorage<RequestContext>();
 
 export class RequestContext {
-  public cookiesWrapper: ReturnType<typeof createCookiesWrapper>;
+  public nextCookies: ReturnType<typeof createNextCookies>;
 
   // TODO: multiple revlidation paths
   public revalidate?: RevalidationType;
 
   constructor(public requestHeaders: Headers) {
-    this.cookiesWrapper = createCookiesWrapper(requestHeaders);
+    this.nextCookies = createNextCookies(requestHeaders);
   }
 
-  getSetCookie = () => this.cookiesWrapper.getSetCookie();
+  getSetCookie = () => this.nextCookies.toResponseCookies().toString();
 
   run<T>(f: () => T): T {
     return requestContextStorage.run(this, f);
@@ -37,34 +37,9 @@ export function headers() {
 }
 
 export function cookies() {
-  return getRequestContext().cookiesWrapper.cookies;
+  return getRequestContext().nextCookies.cookies;
 }
 
 export function revalidatePath(path: string) {
   getRequestContext().revalidate = path;
-}
-
-// it seems Next.js's cookies API has to track modified response cookies
-// so that mutable `cookies()` allows reading current cookies as well.
-// https://github.com/vercel/next.js/blob/6795597a50112c4f83bf61caf3681e95816da4c9/packages/next/src/server/web/spec-extension/adapters/request-cookies.ts#L25-L33
-// for now, we just diff two cookies at the end to decide final set-cookie header
-function createCookiesWrapper(requestHeaders: Headers) {
-  const requestCookies = new RequestCookies(requestHeaders);
-  const cookies = new ResponseCookies(new Headers());
-  for (const cookie of requestCookies.getAll()) {
-    cookies.set(cookie);
-  }
-
-  function getSetCookie() {
-    const result = new ResponseCookies(new Headers());
-    for (const cookie of cookies.getAll()) {
-      const reqCookie = requestCookies.get(cookie.name);
-      if (reqCookie?.value !== cookie.value) {
-        result.set(cookie);
-      }
-    }
-    return result.toString();
-  }
-
-  return { cookies, getSetCookie };
 }
